+++ app/tests/base.py
@@ -2,8 +2,11 @@
 
 from flask_testing import TestCase
 
-from app import app, db
+from app import create_app, db
 from instance.config import app_config
+
+
+app = create_app()
 
 
 class BaseTestCase(TestCase):